fix(direct-translation): issue with translating relations in components
When translating relations in a component, the relation is not translated and when clicking on the
relation field an error is thrown. The bug was that the relation parsing did not handle repeated
components correctly. Another connected issue was that after translating, all fields in the repeated
components were opening simultaneously, which was caused by a missing __temp_key__.

fix #152